@@ -9,24 +9,23 @@ import de.gurkenlabs.litiengine.entities.IEntity;
 import de.gurkenlabs.litiengine.graphics.RenderType;
 import de.gurkenlabs.litiengine.graphics.emitters.EntityEmitter;
 import de.gurkenlabs.litiengine.graphics.emitters.particles.Particle;
-import de.gurkenlabs.litiengine.graphics.emitters.particles.RectangleFillParticle;
+import de.gurkenlabs.litiengine.graphics.emitters.particles.RectangleParticle;
 import de.gurkenlabs.litiengine.physics.Collision;
-import de.gurkenlabs.litiengine.util.ColorHelper;
 
-@EmitterInfo(emitterTTL = 1000, originAlign = Align.LEFT, originVAlign = Valign.MIDDLE, particleMinTTL = 200, particleMaxTTL = 500, spawnRate = 0, spawnAmount = 20, maxParticles = 80)
+@EmitterInfo(duration = 1000, originAlign = Align.LEFT, originValign = Valign.MIDDLE, particleMinTTL = 200, particleMaxTTL = 500, spawnRate = 0, spawnAmount = 20, maxParticles = 80)
 @EntityInfo(renderType = RenderType.OVERLAY)
 public class WaterSplashEmitter extends EntityEmitter {
 
   public WaterSplashEmitter(IEntity entity) {
-    super(entity,true);
-    this.addParticleColor(ColorHelper.decode("#801054A7"));
-    this.addParticleColor(ColorHelper.decode("#80163963"));
-    this.addParticleColor(ColorHelper.decode("#800791BE"));
+    super(entity, true);
+    data().getColors().add("#801054A7");
+    data().getColors().add("#80163963");
+    data().getColors().add("#800791BE");
   }
 
   @Override
   protected Particle createNewParticle() {
-    final float x = this.getRandomParticleX();
+    final float x = data().getParticleOffsetX().getRandomNumber();
     final float y = Game.random().nextFloat(2, 7);
     final float width = 1;
     final float height = 1;
@@ -37,8 +36,8 @@ public class WaterSplashEmitter extends EntityEmitter {
     final float deltaWidth = Game.random().nextFloat(.2f);
     final float deltaHeight = Game.random().nextFloat(.2f);
 
-    return new RectangleFillParticle(width, height, this.getRandomParticleColor(), this.getRandomParticleTTL())
-        .setDeltaX(dx).setDeltaY(dy).setDeltaIncX(gravityX).setDeltaIncY(gravityY).setX(x).setY(y)
+    return new RectangleParticle(width, height).init(data())
+        .setVelocityX(dx).setVelocityY(dy).setAccelerationX(gravityX).setAccelerationY(gravityY).setX(x).setY(y)
         .setDeltaWidth(deltaWidth).setDeltaHeight(deltaHeight).setCollisionType(Collision.STATIC);
   }
 
